Criada a proteção de rotas e contexto de usuário
 Criando o contexto de usuário e autenticação, alterado para uma página de login apenas, rotas protegidas podendo ser acessadas apenas com o usuário logado
