+++ Hooks/UseAuthContext.ts
@@ -1,0 +1,6 @@
+import { useContext } from "react";
+import { AuthContext } from "@/Context/AuthContext";
+
+export default function useAuthContext (){
+    return useContext(AuthContext)
+}
