+++ Model/User.ts
@@ -1,0 +1,7 @@
+export default interface User{
+    id: number;
+    nome: string;
+    email: string;
+    role: string;
+
+}
