+++ package.json
@@ -17,11 +17,13 @@
     "autoprefixer": "^10.4.20",
     "cookie": "^1.0.2",
     "cookies-next": "^5.0.2",
+    "js-cookie": "^3.0.5",
     "next": "15.0.3",
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
   },
   "devDependencies": {
+    "@types/js-cookie": "^3.0.6",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
