+++ pages/_app.tsx
@@ -1,6 +1,13 @@
+import { AuthProvider } from "@/Context/AuthContext";
 import "@/styles/globals.css";
 import type { AppProps } from "next/app";
 
 export default function App({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />;
+ 
+  return(
+   <AuthProvider>
+     <Component {...pageProps} />;
+   </AuthProvider>
+  )
+
 }
